pyproject.toml
[project]
name = "pal-router"
version = "0.1.0"
description = "Ternary LLM router addressing the reasoning cliff"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "textstat>=0.7.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "routellm>=0.1.0",
    "google-generativeai>=0.5.0",
]

[project.optional-dependencies]
eval = ["pandas>=2.0.0", "matplotlib>=3.7.0"]
dev = ["pytest>=7.0.0", "ruff>=0.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pal_router"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
